refactor(convex): consolidate common validators into shared module
Extract duplicated validators into a shared common module to reduce
maintenance overhead and ensure consistency across models.

Changes:
- Create model/common/validators.ts with sortOrderValidator,
  priorityValidator, and dataSourceValidator
- Update 7 model validator files to import from common module
- Maintain backwards compatibility via re-exports

Closes #51
