+++ services/platform/convex/model/conversations/validators.ts
@@ -3,6 +3,8 @@
  */
 
 import { v } from 'convex/values';
+
+import { priorityValidator } from '../common/validators';
 import { approvalItemValidator } from '../approvals/validators';
 
 /**
@@ -16,14 +18,9 @@ export const conversationStatusValidator = v.union(
 );
 
 /**
- * Conversation priority validator
+ * Conversation priority validator (alias for priorityValidator)
  */
-export const conversationPriorityValidator = v.union(
-  v.literal('low'),
-  v.literal('medium'),
-  v.literal('high'),
-  v.literal('urgent'),
-);
+export const conversationPriorityValidator = priorityValidator;
 
 /**
  * Message status validator
